Add user parent class
